api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
playwright==1.48.0
pydantic==2.9.2
openai==1.54.0
python-dotenv==1.0.1

